services/_template/requirements.txt
# Service Name - Requirements
# Replace this header with your service name

# Include shared components
-r ../../shared/requirements.txt

# Web framework (choose one based on your service type)
fastapi>=0.104.0        # For REST API services
uvicorn>=0.24.0        # ASGI server for FastAPI

# OR for background workers, you might prefer:
# celery>=5.3.0         # For task queue workers
# dramatiq>=1.15.0      # Alternative task queue

# Service-specific dependencies (examples - customize for your service)
python-multipart>=0.0.6    # For file uploads (API services)
python-jose[cryptography]>=3.3.0  # For JWT tokens (API services)
passlib[bcrypt]>=1.7.4     # For password hashing (API services)

# Monitoring and observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0  # If using FastAPI

# Development dependencies (add to requirements-dev.txt if you prefer)
pytest-httpx>=0.21.0      # For testing HTTP clients
pytest-mock>=3.12.0       # For mocking in tests
factory-boy>=3.3.0        # For test data generation

# Add your service-specific dependencies below:
# examples:
# scikit-learn>=1.3.0      # For ML services
# transformers>=4.35.0     # For NLP services
# opencv-python>=4.8.0     # For image processing
# sqlalchemy>=2.0.0        # For database services